tools/voice_chat/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
anthropic==0.25.1
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
